+++ packages/variance/integration/__tests__/variance-test.ts
@@ -1,6 +1,5 @@
-import { variance } from '../src/core';
-import { transformSize } from '../src/transforms/transformSize';
-import { theme } from './__fixtures__/theme';
+import { variance } from '../../src/core';
+import { transformSize } from '../../src/transforms/transformSize';
 
 const space = variance.create({
   margin: { property: 'margin', scale: 'spacing' },
