+++ packages/variance/integration/index.ts
@@ -1,0 +1,7 @@
+import { BaseTheme } from '../src/types/theme';
+
+type LocalTheme = typeof theme;
+
+declare module '@emotion/react' {
+  export interface Theme extends BaseTheme, LocalTheme {}
+}
