+++ packages/variance/tsconfig.json
@@ -3,5 +3,6 @@
   "compilerOptions": {
     "outDir": "dist"
   },
+  "exclude": ["./integration"],
   "references": [{ "path": "./src" }]
 }
